+++ src/utils/LocaleUtils.js
@@ -1,5 +1,6 @@
 // Node.js imports
 import { writeFile } from 'node:fs/promises'
+import { existsSync, readFileSync } from 'node:fs'
 import { join } from 'node:path'
 
 /**
@@ -15,7 +16,14 @@ export async function writeLocales (locales, localesFolder) {
   for (const locale of Object.keys(locales)) {
     const localeFilename = join(localesFolder, `${locale}.json`)
     const localeJsonString = JSON.stringify(locales[locale], null, 2)
-    promises.push(writeFile(localeFilename, localeJsonString))
+    if (existsSync(localeFilename)) {
+      const fileContents = readFileSync(localeFilename, 'utf8')
+      if (localeJsonString !== fileContents) {
+        promises.push(writeFile(localeFilename, localeJsonString))
+      }
+    } else {
+      promises.push(writeFile(localeFilename, localeJsonString))
+    }
   }
 
   return Promise.all(promises)
